Move wipe_blockdevs into blockdevmgr module
Once a device is a blockdev it is not really raw any more. Note that it
calls StratBlockDev::disown, which itself must call into methods on
the raw device.
