+++ src/engine/strat_engine/raw_devices/mod.rs
@@ -9,6 +9,6 @@ mod udev;
 
 pub use self::{
     crypt::CryptHandle,
-    devices::{initialize_devices, process_and_verify_devices, wipe_blockdevs},
+    devices::{initialize_devices, process_and_verify_devices},
     identify::{find_all, identify_block_device, DeviceInfo, LuksInfo, StratisInfo},
 };
